+++ src/Renderer/Components/MeshRendererComponentBase.hpp
@@ -19,29 +19,29 @@ namespace Darius::Graphics
 	public:
 
 		// States
-		virtual void						Awake() override;
-		virtual void						OnDestroy() override;
+		virtual void							Awake() override;
+		virtual void							OnDestroy() override;
 
-		virtual void						OnDeserialized() override;
+		virtual void							OnDeserialized() override;
 
 #ifdef _D_EDITOR
-		virtual bool						DrawDetails(float params[]) override;
+		virtual bool							DrawDetails(float params[]) override;
 #endif
 
-		bool								AddRenderItems(std::function<void(D_RENDERER_FRAME_RESOURCE::RenderItem const&)> appendFunction);
+		bool									AddRenderItems(std::function<void(D_RENDERER_FRAME_RESOURCE::RenderItem const&)> appendFunction);
 		
-		INLINE virtual UINT					GetNumberOfSubmeshes() const { return 0; }
+		INLINE virtual UINT						GetNumberOfSubmeshes() const { return 0; }
 
-		virtual bool						CanRender() const { return IsActive(); }
+		virtual bool							CanRender() const { return IsActive(); }
 
 		INLINE virtual D_MATH_BOUNDS::BoundingSphere const& GetBounds() const { return (D_MATH_BOUNDS::BoundingSphere&)*this; }
 
-		INLINE D3D12_GPU_VIRTUAL_ADDRESS	GetConstantsAddress() const { return mMeshConstantsGPU.GetGpuVirtualAddress(); }
+		INLINE D3D12_GPU_VIRTUAL_ADDRESS		GetConstantsAddress() const { return mMeshConstantsGPU.GetGpuVirtualAddress(); }
 
 	protected:
 
 		void									_SetMaterial(UINT index, D_RESOURCE::ResourceHandle handle);
-		uint16_t								GetPsoIndex(UINT materialIndex);
+		UINT									GetPsoIndex(UINT materialIndex);
 		void									OnMeshChanged();
 		
 		DField(Serialize)
@@ -58,11 +58,11 @@ namespace Darius::Graphics
 
 		struct MaterialPsoData
 		{
-			uint16_t								CachedMaterialPsoFlags = 0;
-			uint16_t								PsoIndex = 0;
-			bool									PsoIndexDirty = true;
+			uint16_t							CachedMaterialPsoFlags = 0;
+			UINT								PsoIndex = 0;
+			bool								PsoIndexDirty = true;
 		};
-		D_CONTAINERS::DVector<MaterialPsoData>		mMaterialPsoData;
+		D_CONTAINERS::DVector<MaterialPsoData>	mMaterialPsoData;
 		
 
 	public:
